refactor: Refacto compute similarities and clutering
