--- jobs/ml_jobs/event_linkage/tests/hello_test.py
@@ -1,2 +1,0 @@
-def test_sample():
-    assert int(1.5) == 1
